🐛 Fix list deletion presence transformation
This change fixes a bug where deleting items in a list ahead of your
presence won't correctly transform the presence.

This happens because we change the `ld` to an `oi` to get the correct
behaviour when deleting the actual list item your presence is on.

In order to handle both cases, this change adds a check for the path.
If the path matches the component, then we change it into an `oi` for
the special behaviour, otherwise we leave it so that the `transform()`
works properly.