🐛 No-op `left` transforms where the inserted objects are identical
At the moment if you transform two identical `oi` ops by one another,
you get a no-op in the `right` case, but the `left` case results in a
"silly" op whose `od` value is identical to its `oi` value.

`left` vs `right` should simply determine which op wins in a tie. Since
the two ops are identical, the `right` scenario and `left` scenario
should also be identical.

This change updates the transform behaviour to check this case, and
results in a no-op regardless of `left` or `right` if the two ops are
performing identical `oi`.